From: Tai, Yu-Tsung Sent: Sunday, May 01, 2016 18:45 To: Ortiz, Gerardo <ortizg@indiana.edu>; Hanson, Andrew J. <hansona@indiana.edu>; Sabry, Amr A. <sabry@indiana.edu> Subject: RE: Interval-valued probability
Besides, about the Meyer-Mermin debate, these two follow-up papers may
be interesting to read:

[6] Robert W Spekkens. Contextuality for preparations, transformations,
and unsharp measurements. Physical Review A, 71(5):052108, 2005.
As Gerardo said before, experimentalists in Perimeter Institute tried to
experimental verifying contextuality, but they found the K-S theorem
cannot be directly experimentally verified. Therefore, Spekkens proposed
a new definition of contextuality in this paper, and maybe they are
experimental verifying this new definition?

[7] Otfried Gühne, Matthias Kleinmann, Adán Cabello, Jan-Åke Larsson,
Gerhard Kirchmair, Florian Zähringer, Rene Gerritsma, and Christian F
Roos. Compatibility and noncontextuality for sequential measurements.
Physical Review A, 81(2):022121, 2010.
This paper addressed the contextuality, and whether commuting
observables can measured simultaneously or not in an experiment. We may
need to cite this if we want to said commuting observables may not be
compatible in DQT.

I found these two papers when I wrote my literature review homework, but
I haven’t gone through the details… It seems appropriate to go through
their details if we want to address Meyer-Mermin debate in our paper.

Best regards, 
Tai, Yu-Tsung (戴淯琮)
PhD student in Mathematics & Computer Science
Indiana University, Bloomington

From: Ortiz, Gerardo 
Sent: Sunday, May 01, 2016 16:56
To: Hanson, Andrew J. <hansona@indiana.edu>; Sabry, Amr A.
<sabry@indiana.edu>
Cc: Tai, Yu-Tsung <yuttai@indiana.edu>
Subject: Re: Interval-valued probability

I like very much the way things are unfolding. I thought something
similar to what 
Andy mentioned below. Can one add "topology" to the probability
measure's "range"?
This seems natural given the philosophy we have been advocating in our
Discrete 
Quantum Theories. However, one may argue that the "range" represents the
observer's resources and there is nothing in the mind of the observer
allowing 
for modular arithmetic. I do not see why the observer should wrap
around. Probability 
is a measure of the lack of knowledge (at least at the classical level),
and there is no 
a priori reason to think that it has any particular topology. The fact
that I do not see 
now does not mean that he/she does not use a non-trivial topology to
assign discrete 
probabilities. 

    Best,

      Gerardo 



On 5/1/16 10:44 AM, Hanson, Andrew J. wrote:
I think the most important part is the generalization of partition of
unity. The "sum = 1" part of traditional probability, which we adapted
to Bloch in DQT1 , is what we replaced by the TOPOLOGICAL 'or' operation
this week.  One needs the extension of partition of unity to
Fp-compatible sets instead of the unit interval, and that should embody
the extension of Born's rule and probability measures to quantum theory
based on Fp^2?
  Andy
Sent from my iPhone 5

On May 1, 2016, at 6:19 AM, Sabry, Amr A. <sabry@indiana.edu> wrote:

Formally I think what we want is: 

given a “fuzzy” probability measure mapping “fuzzy events” to interval
probabilities can we reconstruct a “fuzzy state” from which this measure
was calculated using the Born rule. This would be a generalization of
the Born rule and I think this corresponds to Mermin’s response. 

—Amr

P.S. By the way there is a ton of references on “fuzzy quantum” this or
that and some of them touch on the above but I didn’t find anything that
is exactly the above.

On Apr 30, 2016, at 9:41 PM, Sabry, Amr A. <sabry@indiana.edu> wrote:

I thought of Dempster-Shafer before but never made the connection too
:-)

By the way I think we have an immediate paper about the Meyer-Mermin
debate. We can formalize the situation as follows:

* Start with conventional quantum mechanics

* Let states represent a neighborhood. So when we see 'a |0> + b |1>' we
think of all the states within distance delta

* Let observables represent a neighborhood too. So when we ask a
question P we are actually asking about a range of possible questions
[need to formalize this property)

* modify the quantum probability space to be interval-like, which means
modify the Born rule to work with these fuzzy stats and fuzzy questions
to give intervals instead

* Check if Gleason’s theorem now holds

—Amr

On Apr 30, 2016, at 6:14 PM, Hanson, Andrew J. <hansona@indiana.edu>
wrote:

Hi Amr-
 This is embarrassing - I should have recognized the footprints of
Dempster-Shafer here!   When I was at SRI there was a whole group doing
AI expert systems based on this.  I have mostly forgotten the details,
but the principles seem directly adaptable to overlapping interval
quantum probabilities that would avoid all the nogo theorems for DQT.
Need to dig deeper into the formalism: Dan Friedman's student John
Lowrance is an old friend and one of the gurus (with Tom Garvey) of
Dempster-Shafer- I've even met Shafer at SRI a few times.   Tons of
publications by them.  This looks definitely adaptable to our needs. 
Andy

Sent from my iPhone 5

On Apr 30, 2016, at 4:36 PM, Sabry, Amr A. <sabry@indiana.edu> wrote:

Yu-Tsung and I did some good work on interval-valued probability. The
writing is still a bit of a mess but I wanted you to take a look and
think of the following idea which is inspired by how interval-valued
probability is formalized especially as outlined in Sec. 4. 

Let’s go back to the the very beginning of DQC. We were interpreting
states in DQC as sharp but wouldn’t make a lot more sense to think of a
DQC state as an approximation of a collection of actual quantum states.
When the field is small you cluster all the states in two or three
groups. When the field gets bigger you start refining the partitions and
getting more and more accurate. 

Consider F2 again: the states which I am going to call “abstract states"
are

0, |0>, |1>, and |0>+|1>

The abstract state 0 approximates the real state 0

For the other states, think of the Bloch sphere: the quantum states are
all on the surface with |0> at one pole and |1> at the other. If I were
to abstract this, I would make the abstract state |0> correspond to the
real state |0>, the abstract state |1> correspond to the real state |1>,
and every other state of the form alpha|0>+beta|1> with alpha and beta
non-zero is abstracted by |0>+|1>. 

The abstract space would also abstract the unitary evolutions: any
actual evolution between states would correspond to an abstract unitary
in the abstract space and similarly for observables. 

Now the punch line is that this is very similar to Ex.4. The system is
described by approximate information (abstract quantum states) and the
questions we ask are approximate (age range) and hence the probability
are intervals. 

I obviously need to explain this much better but I just wanted to write
it up. —Amr



<prob.pdf>



-- 
*************************************************************************
Gerardo Ortiz
Indiana University                          Phone #: (812) 855-0367
Department of Physics                       Fax   #: (812) 855-5533
727 E. Third St. - Swain West 230           E-mail : ortizg@indiana.edu
Bloomington, IN 47405
*************************************************************************
http://www.iub.edu/~iubphys/research/faculty/Ortiz.shtml
*************************************************************************