From: Sabry, Amr A. Sent: Tuesday, September 27, 2016 9:25 To: Tai, Yu-Tsung <yuttai@indiana.edu>; Ortiz, Gerardo <ortizg@indiana.edu>; Hanson, Andrew J. <hansona@indiana.edu>; Liu, Weihua <liuweih@indiana.edu> Subject: Shapley 1965
Hi,

A central piece of information is the theorem of Shapley 1965 cited in
our draft. A few thoughts on that:

* Administrative: Yu-Tsung, you cite [28,29] for the theorem but only
[28] is by Shapley. Also that paper is an internal memorandum but there
is a published version
http://link.springer.com/article/10.1007/BF01753431
that should be cited instead.

* Yu-Tsung proves the theorem fails in the quantum case. Yesterday, we
came up with a tentative interpretation of that result but I kept
thinking about it. Going back to the original paper, Shapley gives an
interesting intuitive explanation of his theorem. Here is the story:

- We have n players and each player can get a score between [0,1]
- We have coalitions of players that represent subsets (think
projections; subspaces; etc) and they want to collaborate to maximize
their score
- Shapley imposes a condition of convexity that intuitively says that
there is always an incentive to merge, i.e., you can do no worse by
merging a coalition than by playing by yourself.

Under the convexity condition, Shapley proves that every interval-valued
probability measure has a real-valued core.

I question the convexity condition in the quantum case, however. Given
that we have negative probability amplitudes, it is certainly possible
to have situation in which the total probability decreases by joining a
"coalition".

Thoughts?? --Amr