From: Hanson, Andrew J. Sent: Monday, February 06, 2017 17:25 To: Tai, Yu-Tsung <yuttai@indiana.edu> Subject: Need a definition.
Hi Yu-Tsung -
    Looking at the paper, I think you need an improvement:  I would like
to see a one-paragraph, super-clear, definition of "core".
Followed by two examples that show what a core is for each of the
frameworks where you use the terminology.  Perhaps referring to the
2x2 matrix we suggested today for Appendix C's results:

                                 classical     quantum
  smooth probability measure       xxx          yyy

   interval-valued measure         zzz          www


(was that the matrix, or did we have something different?)


   -Andy

  -- if this goes over a page, stop and start over.  I want a very very
clear very succinct but understandable conceptual exposition of what a
core is and why you use the term, and where it comes in of importance.
The non-smooth jumping around of possible interval-outcomes has to
appear naturally, and in principle should tell WHERE there might be an
error in the usage/definition of convexity, both classical and quantum.