From: Tai, Yu-Tsung Sent: Friday, February 10, 2017 0:36 To: Hanson, Andrew J. <hansona@indiana.edu> Cc: Sabry, Amr A. <sabry@indiana.edu>; Ortiz, Gerardo <ortizg@indiana.edu> Subject: RE: Need a definition.
How about something like this?

The paper adopting IOP Guideline is called prop_iop.tex and in
ShareLaTeX as usual...

On Wednesday, Prof. Sabry and I have discussed the idea of understanding
an IVPM as a person's belief. Since it's just a belief, human could have
a consistent or non-consistent belief, and a non-consistent belief might
not be a problem...
In this context, the meaning of "consistent" becomes important.
Classically, a person's belief can be counted as "consistent" if it
satisfies the probability axioms which eliminating the Dutch book.
Similar argument might applied to classical IVPM, and define a version
of Dutch book for IVPM... 
In quantum, if the "consistent" is just eliminating the Dutch book, I
believe a 2-dimensional non-Born probability measure can also eliminate
the Dutch book, and "consistent." My point is: the Born rule requires
more than eliminating the Dutch book. If we want to say a non-Born
probability measure or IVPM is not "consistent," our reason needs to be
more than just the Dutch book, but something else...

Best regards,
Tai, Yu-Tsung (戴淯琮)
PhD student in Mathematics & Computer Science Indiana University,
Bloomington

-----Original Message-----
From: Hanson, Andrew J. 
Sent: Monday, February 06, 2017 17:25
To: Tai, Yu-Tsung <yuttai@indiana.edu>
Subject: Need a definition.

Hi Yu-Tsung -
    Looking at the paper, I think you need an improvement:  I would like
to see a one-paragraph, super-clear, definition of "core".
Followed by two examples that show what a core is for each of the
frameworks where you use the terminology.  Perhaps referring to the
2x2 matrix we suggested today for Appendix C's results:

                                 classical     quantum
  smooth probability measure       xxx          yyy

   interval-valued measure         zzz          www


(was that the matrix, or did we have something different?)


   -Andy

  -- if this goes over a page, stop and start over.  I want a very very
clear very succinct but understandable conceptual exposition of what a
core is and why you use the term, and where it comes in of importance.
The non-smooth jumping around of possible interval-outcomes has to
appear naturally, and in principle should tell WHERE there might be an
error in the usage/definition of convexity, both classical and quantum.