From: Sabry, Amr A. Sent: Monday, April 03, 2017 9:43 To: Tai, Yu-Tsung <yuttai@indiana.edu> Cc: Sabry, Amr A. <sabry@indiana.edu>; Hanson, Andrew J. <hansona@indiana.edu>; Ortiz, Gerardo <ortizg@indiana.edu> Subject: Re: Articles on Stern Gerlach history
https://journals.aps.org/prl/pdf/10.1103/PhysRevLett.101.210401

On Apr 3, 2017, at 9:39 AM, Tai, Yu-Tsung <yuttai@indiana.edu> wrote:

Spekkens' new definition of contextuality...

Tai, Yu-Tsung (戴淯琮)
PhD Candidate in Mathematics & Computer Science
Indiana University, Bloomington

-----Original Message-----
From: Tai, Yu-Tsung
Sent: Monday, April 03, 2017 9:36
To: Hanson, Andrew J. <hansona@indiana.edu>
Cc: Ortiz, Gerardo <ortizg@indiana.edu>; Sabry, Amr A. <sabry@indiana.edu>
Subject: RE: Articles on Stern Gerlach history

We perform a photonic implementation of one such test, ruling out noncontextual models with high confidence.

Tai, Yu-Tsung (戴淯琮)
PhD Candidate in Mathematics & Computer Science Indiana University, Bloomington

-----Original Message-----
From: Hanson, Andrew J.
Sent: Sunday, April 02, 2017 14:22
To: Tai, Yu-Tsung <yuttai@indiana.edu>
Cc: Hanson, Andrew J. <hansona@indiana.edu>; Ortiz, Gerardo <ortizg@indiana.edu>; Sabry, Amr A. <sabry@indiana.edu>
Subject: Articles on Stern Gerlach history

Hi all -
  That paper I sent has some pretty interesting stuff in it.
In particular, all the results are actually incorrectly interpreted, and even off by a factor of two if it were not for the (then not understood) Thomas precession.
(I actually have an interactive program that will visualize the Thomas precession very nicely if anyone is interested.
It's a side effect of an app I used in my computer graphics course to explain rotations, get Thomas just by changing a cosine to a cosh etc.)

   Here is the part I liked:

http://www.fhi-berlin.mpg.de/mp/friedrich/PDFs/ptsg.pdf

Those and other puzzles, such as the anomalous Zeeman effect, could not be cleared up until several years later, after the development of quantum mechanics and the inclusion of electron spin in the theory.
Those advances made the Bohr model obsolete but enhanced the scope and significance of space quantization.  The gratifying agreement of the Stern– Gerlach splitting with the old theory proved to be a lucky coincidence. The orbital angular momentum of the silver atom is actually zero, not hbar/2 pi as presumed in the Bohr model. The magnetic moment is due solely to a half unit of spin angular momentum, which accounts for the twofold splitting. The magnetic moment is nonetheless very nearly one Bohr magneton, by virtue of the Thomas factor of two, not recognized until 1926. Nature thus was duplicitous in an uncanny way.

A curious historical puzzle remains. In view of the interest aroused by the SGE in 1922, we would expect that the postulation of electron spin in 1925 should very soon have led to a reinterpretation of the SGE splitting as really due to spin.  However, the earliest attribution of the splitting to spin that we have found did not appear until 1927, when Ronald Fraser noted that the ground-state orbital angular momentum and associated magnetic moments of silver, hydrogen, and sodium are zero.  Practically all current textbooks describe the Stern–Gerlach splitting as demonstrating electron spin, without pointing out that the intrepid experimenters had no idea it was spin that they had discovered.
<R. Spekkens, Contextuality for preparations, transformations, and unsharp measurements.pdf>