From: Hanson, Andrew J. Sent: Wednesday, May 10, 2017 11:27 To: Tai, Yu-Tsung <yuttai@indiana.edu> Cc: Sabry, Amr A. <sabry@indiana.edu>; Ortiz, Gerardo <ortizg@indiana.edu> Subject: Here is the first draft intro text
Here is our starting draft of the intro page text:
   -- the last paragraphs in particular depend a lot on opinions of what
we should emphasize, and so are heavily open to discussion, revision,
and improvement.
    -Andy and Yu-Tsung
----------------------------------------------------

In this letter, we extend our previous work on the foundations of
computability in quantum mechanics by exploring the application of
interval-valued probability measures (IVPM) to achieving a coherent
formulation of finite-resource quantum mechanics.

There are long-standing debates in the foundations of quantum mechanics
regarding the tension between finite precision measurement and
contextuality~\cite{Peres2003,BarrettKent2004}.  For example,
Meyer~\cite{PhysRevLett.83.3751} and Mermin~\cite{Mermin1999} disagree
on whether finite-precision measurement invalidate the Kochen-Specker
theorem~\cite{kochenspecker1967,peres1995quantum}, and discussion
continues on whether contextuality is experimentally
verifiable~\cite{Spekkens2005,GuehneKleinmannCabelloEtAl2010,MazurekPuseyKunjwalEtAl2016}. 
The
fundamental question is how to make mathematical theories of quantum
mechanics that are intrinsically, rather than only implicitly,
consistent with the resources available for the realistic accuracy of an
actual measurement.

Our first observation is that essential objectives of our program to
achieve a computable theory of quantum mechanics and finite-resource
measurement, attempted previously with computable number
systems~\cite{usat,geometry2013,DQT2014}, may be achievable by extending
classical interval-valued probability measures~\cite{JamisonLodwick2004}
to the quantum domain.  We begin by axiomatizing a quantum
interval-valued probability measure (QIVPM) framework exploiting the
fact that a density matrix approach can be considered as equivalent to a
description of quantum states using probability
measures~\cite{10.2307/2308516,Varadarajan2008}. %Mackey=10.2307 One
result is that the expectation value of an observable with respect to a
QIVPM can be defined to be consistent with the constraints on the
expectation value of the same observable expected from continuous
quantum mechanics.  We then examine Gleason's
theorem~\cite{gleason1957,Redhead1987-REDINA,peres1995quantum} and the
Kochen-Specker theorem in the context of QIVPM's, and conclude that they
correspond to the limit of an infinite number of precise intervals, and
the set of crypto-deterministic intervals (the set
$\{\imposs,\necess\}$), respectively.

We note, however, that in this approach there exist QIVPM's that do not
correspond to a density matrix computable by the Born rule, so that
Gleason's theorem cannot hold in general.  It thus remains possible that
a refinement of the basic definition of a quantum state or the
computation of probability from a state may be required when we
rigorously consider the implications of finite precision measurement.