From: Hanson, Andrew J. Sent: Saturday, December 02, 2017 3:09 To: Tai, Yu-Tsung <yuttai@indiana.edu> Cc: Ortiz, Gerardo <ortizg@indiana.edu>; Sabry, Amr A. <sabry@indiana.edu> Subject: AJH typo corrections and separate major suggested edits: re: AL11679 Tai
Hi all -
    Had another disastrous day at home, ending with a shower repair
turning into gallons of water pouring through the ceiling into the
living room.  If it's not fire, it's floods, it seems.  We are really
not having good luck this week as homeowners!

   I did force myself to put the edits and suggestions I had prepared to
correct the typos for the Tuesday non-meeting with Yu-Tsung, and they
are attached:

   Part 1 is  prop_letter_ajh.tex, which is the "submitted"
version with various typos and non-controversial fixes, each marked by
"%% ajh  <explanation>", so you can find the changes easily by searching
for 'ajh', and also I kept the format of the ShareLatex as close as
possible so  diff or xxdiff  should make it very easy to find what I did
and accept or ignore it.  This was mostly typos.

   Part 2 is prop_SUGGESTED_ajh.txt,  which are the major rewrites of
various sections, noted in the .tex file in the form
    %% ajh  SUGGESTED TEXT and a number.
These are NOT integrated into the ShareLatex .tex file, but noted as
comments with a number that corresponds in the two files.  These are
much bigger corrections, as the file is 1000 words long, representing my
personal preferences and very possibly my personal misunderstandings or
inadequacies in following some of the written discussion sections in the
paper in the form submitted Tuesday.

   -Andy