From: Ortiz, Gerardo Sent: Saturday, December 02, 2017 18:17 To: Tai, Yu-Tsung <yuttai@indiana.edu>; Hanson, Andrew J. <hansona@indiana.edu> Cc: Sabry, Amr A. <sabry@indiana.edu> Subject: Re: AJH typo corrections and separate major suggested edits: re: AL11679 Tai
My suggestions are:

1) Put Andy's prop_letter_ajh.tex in Sharelatex AS A SEPARATE FILE.

2) Produce a NEW file prop_letter_ajhs.tex AS ANOTHER SEPARATE FILE, and
put it ALSO in Sharelatex.

In the end there will be THREE tex files that we need to compare on
Thursday. 

   Thanks,

      Gerardo

PS: If we start mixing files we will screw things again as we apparently
did in that paragraph of the 
submitted version. 


On 12/2/17 6:09 PM, Tai, Yu-Tsung wrote:
Dear all,
 
OK. I would merge Andy’s suggestion into the main version in
GitHub/ShareLaTeX. Just want to confirm: Is it OK to merge the another
file into “Andy’s say” boxes? So that it might be easier to see the
changes and comments in one PDF file instead of looking up and comparing
different source files?
 
Best regards,
Tai, Yu-Tsung (戴淯琮)
PhD Candidate in Mathematics & Computer Science
Indiana University, Bloomington
https://www.linkedin.com/in/yu-tsung-tai-9aa30551
 
From: Ortiz, Gerardo 
Sent: Saturday, December 02, 2017 17:51
To: Hanson, Andrew J. <hansona@indiana.edu>; Tai, Yu-Tsung
<yuttai@indiana.edu>
Cc: Sabry, Amr A. <sabry@indiana.edu>
Subject: Re: AJH typo corrections and separate major suggested edits:
re: AL11679 Tai
 
Hi Andy,

   I am sorry to hear you had problems again with your house. Hope you
solve them soon. 
Thanks for the updates on the text. I would suggest that  we discuss 
altogether your prop_letter_ajh.tex and you more extended proposed 
changes next Thursday. I just read your prop_letter_ajh.tex and I
believe that 
the paragraph we screwed in the submission is not correct yet. My
suggestion is 
that YT keeps an updated version that we keep to replace when we receive
the referee 
reports from PRA.

   Best,

     Gerardo


On 12/2/17 3:08 AM, Andrew Hanson wrote:
Hi all - 
   Had another disastrous day at home, ending with a 
shower repair turning into gallons of water pouring 
through the ceiling into the living room.  If it's 
not fire, it's floods, it seems.  We are 
really not having good luck this week as homeowners! 

  I did force myself to put the edits and suggestions 
I had prepared to correct the typos for the Tuesday 
non-meeting with Yu-Tsung, and they are attached: 

  Part 1 is  prop_letter_ajh.tex, which is the "submitted" 
version with various typos and non-controversial fixes, 
each marked by  "%% ajh  <explanation>", so you can find 
the changes easily by searching for 'ajh', and also I 
kept the format of the ShareLatex as close as possible 
so  diff or xxdiff  should make it very easy to find 
what I did and accept or ignore it.  This was mostly 
typos. 


  Part 2 is prop_SUGGESTED_ajh.txt,  which are the major 
rewrites of various sections, noted in the .tex file 
in the form 
   %% ajh  SUGGESTED TEXT and a number. 
These are NOT integrated into the ShareLatex .tex file, 
but noted as comments with a number that corresponds in the 
two files.  These are much bigger corrections, as the 
file is 1000 words long, representing my personal 
preferences and very possibly my personal misunderstandings 
or inadequacies in following some of the written 
discussion sections in the paper in the form submitted 
Tuesday. 

  -Andy 
 
-- 
********************************************************************************
Gerardo Ortiz 	 
Indiana University	Phone #: (812) 855-0367
Department of Physics	Fax #: (812) 855-5533
727 E. Third St. - Swain West 230	E-mail : ortizg@indiana.edu

Bloomington, IN 47405	 
Faculty Webpage
 
Paperback Edition from Oxford University Press  
********************************************************************************

-- 
********************************************************************************
Gerardo Ortiz 	 
Indiana University	Phone #: (812) 855-0367
Department of Physics	Fax #: (812) 855-5533
727 E. Third St. - Swain West 230	E-mail : ortizg@indiana.edu

Bloomington, IN 47405	 
Faculty Webpage
 
Paperback Edition from Oxford University Press  
**************************************************