From: e-prints@arxiv.org [mailto:e-prints@arxiv.org] Sent: Monday, December 25, 2017 12:54 To: Tai, Yu-Tsung <yuttai@indiana.edu> Subject: arXiv submission submit/2114094
We have received your submission to arXiv. Your temporary submission
identifier is: submit/2114094.

You may update your submission at: https://arxiv.org/submit/2114094

Your article is scheduled to be announced at Wed, 27 Dec 2017 01:00:00
GMT.
The abstract will appear in the subsequent mailing as displayed below,
except that the submission identifier will be replaced by the official
arXiv identifier. Updates before Tue, 26 Dec 2017 19:00:00 GMT will not
delay announcement.

A paper password will be emailed to you when the article is announced. 
You should share this with co-authors to allow them to claim ownership.

If you have a problem that you are not able to resolve through the web
interface, contact help@arxiv.org with a description of the issue and
reference the submission identifier.

arXiv admin

------------------------------------------------------------------------------
\\
arXiv:submit/2114094
From: Yu-Tsung Tai <yuttai@indiana.edu>
Date: Sun, 24 Dec 2017 23:53:59 EST   (30kb,D)

Title: Quantum Interval-Valued Probability: Contextuality and the Born
Rule
Authors: Yu-Tsung Tai, Andrew J. Hanson, Gerardo Ortiz, and Amr Sabry
Categories: quant-ph math-ph math.MP
Comments: Submitted to Physical Review A
MSC-class: 81P10, 81P15, 81P16, 91A12
License: http://arxiv.org/licenses/nonexclusive-distrib/1.0/
\\
  We present a mathematical framework based on quantum interval-valued
probability measures to study the effect of experimental imperfections
and finite precision measurements on defining aspects of quantum
mechanics such as contextuality and the Born rule. While foundational
results such as the Kochen-Specker and Gleason theorems are valid in the
context of infinite precision, they fail to hold in general in a world
with limited resources. Here we employ an interval-valued framework to
establish bounds on the validity of those theorems in realistic
experimental environments. In this way, not only can we quantify the
idea of finite-precision measurement within our theory, but we can also
suggest a possible resolution of the Meyer-Mermin debate on the impact
of finite-precision measurement on the Kochen-Specker theorem.
\\