https://math.indiana.edu/graduate/index.html
IU Thesis Document Class

The iuthesis document class is a LaTeX document class that produces output acceptable (as of Summer 2013) for an Indiana University Bloomington Masters or Doctoral thesis. It is available free for anybody's use, but comes with NO WARRANTY. In particular, the author is not responsible if it produces output that is not acceptable to the graduate school. This document class is not officially sanctioned by the University Graduate School, although several theses have been written with it and accepted.

Per the University Graduate School information page on thesis formatting, the template font in the sample is not the correct size – Throughout the dissertation, only size 11 or 12 point font should be used with the only exceptions being the title on the title page and any footnotes. The title, only, on the title page may be as large as 16 point font while footnotes can be no smaller than 10 point. All other font must be size 11 or 12 point font. This requirement includes all content including section and chapter headings. In addition, the Acceptance and Abstract pages are slightly different – official formatting should have lines for the committee signatures.

The Files:
•iuphd.cls
•iuphd-doc.tex
•dissertation.tex
•dissertation.bib
•Example file, also doubles as documentation.